Core-1973 more changes in Menu controller

+++ api/src/main/java/org/ednovo/gooru/infrastructure/persistence/hibernate/menu/MenuRepository.java
@@ -19,7 +19,7 @@ public interface MenuRepository extends BaseRepository{
 	
 	MenuItem findMenuItemById(String menuItemUid);
 	
-	Long getMenuItemCount( String menuUid);
+	Integer getMenuItemCount( String menuUid);
 	
     List<MenuItem> getMenuItemsByMenuId(String menuUid);
